backend-python/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23

# ML & NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP Specific
nltk>=3.8.0
# spacy>=3.7.0  # Optional - install separately if needed

# AI API
google-generativeai==0.3.2

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.3
playwright==1.41.2
googlesearch-python==1.2.5

# OCR & Document Processing
pytesseract>=0.3.10
Pillow>=10.3.0  # Updated for Python 3.13
PyPDF2>=3.0.0
pdf2image>=1.17.0

# Validation
pydantic>=2.0.0
email-validator==2.1.0

# Email Services
aiosmtplib==3.0.1  # Async SMTP for sending emails
jinja2==3.1.3      # Email template rendering

# Utils
python-dateutil==2.8.2
pytz==2024.1

# Monitoring & Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9

# File handling
aiofiles==23.2.1


# Vector Store & RAG
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.12.0
tiktoken>=0.5.2
faiss-cpu>=1.7.4
